Add ogip.minimium_energy config for ARFs
ARFs are now checked to make sure the ENERG_LO/HI columns meet some basic
criteria, based on the OGIP standard (ENERG_HI > ENERG_LO, no negative
values, and that the arrays are monotonically increasing or decreasing).
If the checks fail then a DataErr is raised (this adds a new "catch-all"
DataErr case called 'ogip-error').

The case when the minimum ENERG_LO value is 0 is treated specially, since
although it is against the OGIP standard, there are responses out there
that use a minimum value of 0 for the lower edge of the first bin. There
are two options:

  - error out with a DataErr
  - replace the 0 value

The choice depends on the emin parameter of the DataARF constructor: None
means error, a positive value is used as the replacement value. This is
only intended for the object API: the sherpa.astro.ui layer (actually, the
sherpa.astro.io module) uses a new configuration variable called
minimum_energy in the newly-added ogip block) to set this value. The
default setting is 1e-10.

XSPEC does something similar, but uses a different value.

Future commits will extend this to the RMF case.